requirements.txt
google-generativeai
numpy
matplotlib
python-dotenv
pytest